requirements.txt
premai
python-dotenv